Move InteractionCard to common component folder
