+++ src/Patient/Information/Panel.js
@@ -9,7 +9,6 @@ const useStyles = makeStyles({
     }
 })
 
-
 export default function Test(props){
 
     const classes = useStyles();
