+++ src/Patient/Progress/Milestones.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { makeStyles } from '@material-ui/core/styles';
-import InteractionCard from '../Home/InteractionCard';
+import InteractionCard from '../../Basics/InteractionCard';
 import Styles from '../../Basics/Styles';
 import { DateTime } from 'luxon';
 import Colors from '../../Basics/Colors';
